oiio-sys: move header files next to the implementation files
Make the sources easier to navigate by cohabitating the headers and .cpp sources.
