+++ oiio-sys/src/deepdata.rs
@@ -3,7 +3,7 @@ pub use ffi::*;
 #[cxx::bridge(namespace = oiio)]
 mod ffi {
     unsafe extern "C++" {
-        include!("oiio-sys/include/ffi_deepdata.h");
+        include!("oiio-sys/src/ffi_deepdata.h");
 
         pub type DeepData;
     }
