+++ oiio-sys/src/ffi_deepdata.cpp
@@ -1,4 +1,4 @@
-#include "oiio-sys/include/ffi_deepdata.h"
+#include "ffi_deepdata.h"
 
 namespace oiio {
 }  // namespace oiio
