+++ oiio-sys/src/ffi_filesystem.cpp
@@ -1,4 +1,4 @@
-#include "oiio-sys/include/ffi_filesystem.h"
+#include "ffi_filesystem.h"
 
 namespace oiio {
 }  // namespace oiio
