+++ oiio-sys/src/ffi_imageio.cpp
@@ -1,4 +1,4 @@
-#include "oiio-sys/include/ffi_imageio.h"
+#include "ffi_imageio.h"
 #include "oiio-sys/src/imageio.rs.h"
 #include <OpenImageIO/imageio.h>
 #include <OpenImageIO/string_view.h>
