+++ oiio-sys/src/ffi_imageio.h
