+++ oiio-sys/src/ffi_typedesc.h
