+++ oiio-sys/src/filesystem.rs
@@ -3,7 +3,7 @@ pub use ffi::*;
 #[cxx::bridge(namespace = oiio)]
 mod ffi {
     unsafe extern "C++" {
-        include!("oiio-sys/include/ffi_filesystem.h");
+        include!("oiio-sys/src/ffi_filesystem.h");
 
         pub type IOProxy;
     }
